packages/gnss-product-management/pyproject.toml
[project]
name = "gnss-product-management"
requires-python = ">=3.10,<3.15"
dynamic = ["version"]
dependencies = [
    "aiohttp>=3.13.3",
    "cloudpathlib[s3]>=0.18",
    "fsspec>=2024.0",
    "gnss-management-specs",
    "julian>=0.14",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "requests>=2.32.5",
]

[tool.uv.sources]
gnss-management-specs = { workspace = true }

[dependency-groups]
dev = [
    "pytest>=8.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v -s"
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(levelname)-8s %(name)s — %(message)s"
markers = [
    "integration: marks tests that require live FTP connections (skip with -m 'not integration')",
]

[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "vcs"
raw-options = { root = "../.." }

[tool.hatch.build.targets.wheel]
force-include = { "pyproject.toml" = "pyproject.toml" }

